Display building classifications on titles
