+++ packages/web/src/types/Form.ts
@@ -1,0 +1,12 @@
+interface Option {
+  value: string;
+  label: string;
+  description: string;
+  subOptions?: Array<Option>;
+}
+
+interface ParentLinkedOption extends Option {
+  parent: Option | null;
+}
+
+export type { Option, ParentLinkedOption };
