+++ packages/web/src/types/Title.ts
@@ -1,0 +1,7 @@
+import type { Option } from "./Form";
+
+type LandClassifications = Array<Option>;
+
+type BuildingClassifications = Array<Option>;
+
+export type { LandClassifications, BuildingClassifications };
